refactored compile.jl using compileutils.jl, removed const keyword, deleted particles/ants .jl files
