+++ src/Autumn/Autumn.jl
@@ -17,6 +17,9 @@ include("subexpr.jl")
 include("sexpr.jl")
 @reexport using .SExpr
 
+include("compileutils.jl")
+@reexport using .CompileUtils
+
 include("compile.jl")
 @reexport using .Compile
