+++ src/Autumn/aprograms/ants.au
@@ -1,4 +1,4 @@
-const GRID_SIZE = 16
+GRID_SIZE = 16
 
 type alias Position = { x : Int, y : Int}
 type alias Particle = { position : Position, color : String, render : Boolean }
