+++ src/Autumn/aprograms/particles.sx
@@ -1,5 +1,5 @@
 (program
-  (const (= GRID_SIZE 16))
+  (= GRID_SIZE 16)
 
   (type alias Position ((: x Int) (: y Int)))
   (type alias Particle ((: position Position)))
