+++ src/Autumn/sexpr.jl
@@ -45,7 +45,6 @@ function parseau(sexpr::AbstractArray)
     [:(=), x::Symbol, y]              => AExpr(:assign, x, parseau(y))
     [:(:), v::Symbol, τ]              => AExpr(:typedecl, v, parsetypeau(τ))
     [:external, tdef]                 => AExpr(:external, parseau(tdef))
-    [:const, assignment]              => AExpr(:const, parseau(assignment))
     [:let, vars, val]                 => AExpr(:let, map(parseau, vars)..., parseau(val))
     [:case, name, cases...]           => AExpr(:case, name, map(parseau, cases)...)
     [:(=>), type, value]              => AExpr(:casevalue, parseau(type), parseau(value))
